+++ unstoppable-ios-app/domains-manager-ios/Services/DomainProfilesService/Network/DomainProfileNetworkServiceProtocol.swift
@@ -1,5 +1,5 @@
 //
-//  PublicDomainProfileNetworkServiceProtocol.swift
+//  DomainProfileNetworkServiceProtocol {.swift
 //  domains-manager-ios
 //
 //  Created by Oleg Kuplin on 05.03.2024.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-protocol PublicDomainProfileNetworkServiceProtocol {
+protocol DomainProfileNetworkServiceProtocol {
     func fetchPublicProfile(for domainName: DomainName, fields: Set<GetDomainProfileField>) async throws -> SerializedPublicDomainProfile
     @discardableResult
     func updateUserDomainProfile(for domain: DomainItem,
