+++ unstoppable-ios-app/domains-manager-ios/Services/DomainProfilesService/Storage/DomainProfileDisplayInfoStorageServiceProtocol.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-protocol PublicDomainProfileDisplayInfoStorageServiceProtocol {
+protocol DomainProfileDisplayInfoStorageServiceProtocol {
     func store(profile: DomainProfileDisplayInfo)
     func retrieveProfileFor(domainName: DomainName) throws -> DomainProfileDisplayInfo
 }
